[FEATURE] Include rootpage id in unique alias cache
Different records with the same generated alias get a "-1"
added, also if they are used in different trees only.
This is not necessary and could be avoided by adding
the rootpage id to the unique alias cache. So the aliases
will get a "-1" for unification only if the alias is conflicting
in the same tree.